apps/backend/requirements.txt
gunicorn==23.0.0
Django==5.2
django-environ==0.12.0
django-ckeditor==6.7.2
django-treenode==0.23.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
pillow==11.2.1
psycopg[binary]
